TCP Client-Server App
In this project I have implemented a TCP Client-Server application using the socket and thread libraries. 

Acctually I created a group chat environment, which can be used by users in many options:
1. Users can create their own group chats.
2. New clients can join an existing group chat if they know the correct name and ID of the group. 
3. Clients also have the ability to disconnect from the server and leave the group chat if desired.

To use the application, first, you need to run the server. 
Once the server is up and running, you can launch as many clients as needed to join the group chat.